muda estrutura de pastas e adicionar autenticacao
