+++ AJuDE/src/main/java/psoft/ufcg/api/AJuDE/AJuDeApplication.java
@@ -1,0 +1,13 @@
+package psoft.ufcg.api.AJuDE;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AJuDeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AJuDeApplication.class, args);
+	}
+
+}
