+++ AJuDE/src/main/resources/application.properties
@@ -1,0 +1,17 @@
+# H2
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2
+#indicará o path para você acessar a interface do h2, em geral, vá ao browser e coloque localhost:8080/h2 com 8080 ou sua porta
+
+#deixa que hibernate gerencia a criação das bases de dados - permite atualizações nas bases, mas nunca apaga tabelas ou colunas que não estejam em uso pela aplicação - existem outras configurações - essa é só simples e segura na fase de desenvolvimento!
+spring.jpa.hibernate.ddl-auto=update
+
+# Datasource
+spring.datasource.url=jdbc:h2:file:./dados
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.driver-class-name=org.h2.Driver
+
+ security.jwt.token.expire-length=2520000000 
+
+server.servlet.context-path=/api/v1
