+++ AJuDE/src/test/java/psoft/ufcg/api/AJuDE/AJuDeApplicationTests.java
@@ -1,0 +1,13 @@
+package psoft.ufcg.api.AJuDE;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AJuDeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
