+++ README.md
@@ -1,1 +1,12 @@
-# psoft-backend
+# AJuDE API
+> Projeto backend de PSOFT
+
+## Como rodar
+
+`./mvnw spring-boot:run`
+
+## Endpoints
+
+ Endpoint   |     Método      |  Descrição |
+|----------|:----------:|----------|
+| `/v1/api/campanha/` |  **GET** | Retorna todas as campanhas |
