--- projeto/src/main/java/psoft/projeto/ProjetoApplication.java
@@ -1,13 +1,0 @@
-package psoft.projeto;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProjetoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProjetoApplication.class, args);
-	}
-
-}
