--- projeto/src/test/java/psoft/projeto/ProjetoApplicationTests.java
@@ -1,13 +1,0 @@
-package psoft.projeto;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjetoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
